// src/app/api/interfaces/Examination/studentseatingplaninterface.ts
export interface SeatingPlan {
    Vid: string
    CourseCode: string
    CourseName: string
    ExamTypeDesc: string
    ExamCategory: string
    PaperType: string
    RoomNo: string
    ReportingTime: string
    ExamDate: string
    ExamTime: string
    DetaineeStatus: string
    DefaulterDetail: string
    Instructions: string
    Filters: string
  }